resatructured collisioncheckservice and moved out projectioninterval
